backend/get-auth-token/requirements.txt
functions-framework==3.*
google-auth==2.*
google-auth-httplib2==0.1.*
google-auth-oauthlib==1.*
firebase-admin==6.*
flask==2.*
flask-cors==3.*
google-cloud-core==2.*
